refactor: 기존에 존재하던 예외를 재사용하도록 변경
