--- jdbc/src/main/java/org/springframework/jdbc/core/StatementAgentServiceFailedException.java
@@ -1,8 +1,0 @@
-package org.springframework.jdbc.core;
-
-public class StatementAgentServiceFailedException extends RuntimeException {
-
-    public StatementAgentServiceFailedException(Throwable e) {
-        super(e);
-    }
-}
