+++ jdbc/src/main/java/org/springframework/jdbc/datasource/CannotGetJdbcConnectionException.java
@@ -1,4 +1,4 @@
-package org.springframework.jdbc;
+package org.springframework.jdbc.datasource;
 
 import java.sql.SQLException;
